add /roll
